Run publish-python to test release workflow (#5)
On several occasions, changes to the `setup.cfg` of a colcon package
have resulted in the Debian patch failing to apply. To avoid that in the
future, we should run `publish-python` to validate the release workflow
as part of CI.

Since the Debian part of that workflow will only work on Debian or
Ubuntu, only run that part there.