Options to specify strategy
Adds new arguments the `pytest.yaml` reusable workflow, to make it
easier to specify a strategy. To generate the strategy, a small utility
`genstrat.py` has been added. By default, this utility will output a
file identical to the old `strategy.json` (which has been deleted). It
is still possible to use `matrix-filter`, if a user prefers to use that.

Resolves: #37